Updates Foundation, fix Pear and Packagist support.
- Removes outdated Respect/Foundation Makefile.
- Updates Respect/Foundation Makefile.
- Updates .travis.yml to use Respect/Foundation.
- Bumps version on package.xml for PEAR and cleans composer.json.
  Information removed from composer.json (for Packagist) was not
  necessary.

The *fix* for PEAR and Packagist is not actually a FIX, it it more
like a temporary thing. It does not make use of c9s/Onion's
package.ini file since it does not currently support dependencies
of other packages. Respect/Foundation uses Onion to create
and maintain files: package.xml and composer.json. The real fix for
that should arrive some day soon and them we will have proper
handling of dependencies.